env variables set
