+++ Pipfile
@@ -1,0 +1,19 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+flask = "*"
+"psycopg2-binary" = "*"
+flask-migrate = "*"
+flask-sqlalchemy = "*"
+flask-wtf = "*"
+requests = "*"
+coverage = "*"
+pytest = "*"
+
+[dev-packages]
+
+[requires]
+python_version = "3.6"
