--- src/tests/__init__.py
@@ -1,14 +1,0 @@
-import pytest
-from .. models import db, Company
-
-@pytest.fixture
-def session():
-
-    def do_nothing():
-        pass
-
-    db.session.commit = do_nothing
-
-    yield db.session
-
-    db.session.rollback()
